django-server/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
PyJWT==2.8.0
cloudinary==1.36.0
Pillow==10.1.0
python-decouple==3.8
